+++ libs/common/src/constants/constants.ts
@@ -1,11 +1,3 @@
-export const ERROR_MESSAGES = {
-  INVALID_CREDENTIALS: 'Invalid credentials',
-  USER_NOT_FOUND: 'User not found',
-  USER_ALREADY_EXISTS: 'User already exists',
-  USERNAME_TAKEN: 'Username already taken',
-  CREATE_USER: 'Error creating user',
-  LOGOUT_FAILED: 'Error logging out',
-};
 export const SUCCESS_MESSAGES = {
   USER_CREATED: 'User created successfully',
   USER_LOGGED_OUT: 'User logged out successfully',
