+++ libs/common/src/constants/errors/error.messages.ts
@@ -1,0 +1,8 @@
+export const ERROR_MESSAGES = {
+  INVALID_CREDENTIALS: 'Invalid credentials',
+  USER_NOT_FOUND: 'User not found',
+  USER_ALREADY_EXISTS: 'User already exists',
+  USERNAME_TAKEN: 'Username already taken',
+  CREATE_USER: 'Error creating user',
+  LOGOUT_FAILED: 'Error logging out',
+};
